requirements.txt
pexpect
pyOpenSSL
bugzilla >= 0.8.0
configobj >= 4.7.2
futures
setuptools
jira
lxml
netaddr
paramiko >= 1.15.2
repoze.lru >= 0.6
python-rrmngmnt
python-otopi-mdp
unittest2
pywinrm
PyYAML
winremote
# pytest >= 3.2 before they release it, we have to use our fork.
git+https://github.com/rhevm-qe-automation/pytest@features
pytest-xdist
pytest-ordering
git+git://github.com/rhevm-qe-automation/pytest-catchlog.git@rhevm_customization
# pytest-marker-bugzilla
git+git://github.com/rhevm-qe-automation/pytest-marker-bugzilla.git
uritemplate
ovirt-engine-sdk-python
requests
